Map name and plan from webhook
Without a better guide to go by I've mapped things like so:

name is the human readable name, which I am mapping to the offering.
This isn't very human readable, we may be able to get a human readabe
name from the api but aren't calling it yet.

processor_plan is the processor's plan id, in this case I am picking the
product_id for the unique product behind the offering.

offering: https://www.revenuecat.com/docs/offerings/overview