Find/Create payment processor by user id
In the future we will want to let a developer customise the model and
field name.

We're doing it this way because revenuecat presents a chicen and egg
problem. Because the exchange begins on native code we don't necessarily
have an opportunity to set the payment processor and create a customer.
But we need an ID that we can give to Revenuecat so we can connect the
dots when they send us a webhook.

Possible options for an ID are:

- Pay::Customer.id
We are not guranteed that a user will have a revenuecat payment
processor.

- User.id
Some platforms don't use this model, others will be using accounts
for their transactions, not users.

My personal preference is generating a `public_id` on my model, too.

For now I'm going to use User, and will make it configurable later.