Allow Renewal to handle apple "initial" renewals
Once a subscription is cancelled and expires on Apple, it does not try
to create a new initial purchase if the customer comes back. Instead it
renews the cancelled and expired subscription.

While it may not be a real world condition, when testing we often reset
our databases, and this gets in the way because we won't have the
initial subscription we're expecting, so we need to handle that.

This does mean that we can lose the initial purchase code and only use
the renewal code.