Move INITIAL_PURCHASE to renewal
In the previous commit we had to handle some of the initial purchase
code in the renewal step. There seems little point in maintaining two
versions so now we're making renewal handle both.

Eventually we'll move it onto the subscription class once we're happy
that it's all working correctly. This is how the pay gem works, it has a
`#sync` method on Customer/Subscription/Charge.

Co-Authored-By: Theo Ntoka <uodvraxz@gmail.com>