Add Subscription.resumable?
The view from the pay gem (at least in Jumpstart Pro) expects the
subscription model to respond to .resumable?

I've set it to false because sfaik controlling payments is up to the
user in the ecosystem. If they want to renew they can do so in their
app/play store.